backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
unstructured
pdf2image
pytesseract
pillow
faiss-cpu
sentence_transformers
python-docx
openai
streamlit
requests
pandas
pdfminer.six 
pi-heif
unstructured[all-docs]
python-dotenv
langchain-groq